move model utility functions
