+++ src/model/nodes.ts
@@ -1,8 +1,7 @@
-import { NodeRefLike, NodeRef, INode, IPort, IForwardRefNode, NodeAttributes, IAttributes } from '../types';
+import { INode, IPort, IForwardRefNode, NodeAttributes, IAttributes } from '../types';
 import { DotObject } from './abstract';
 import { attribute } from '../attribute';
 import { Attributes } from './attributes-base';
-import { NodeRefGroupLike } from '../types';
 
 /**
  * Node object.
@@ -24,24 +23,4 @@ export class Node extends DotObject implements INode {
     }
     return { id: this.id, ...port };
   }
-}
-
-/** @hidden */
-export function isForwardRefNode(object: unknown): object is IForwardRefNode {
-  return typeof object == 'object' && object !== null && typeof (object as IForwardRefNode).id === 'string';
-}
-
-/** @hidden */
-export function isEdgeTarget(node: unknown): node is NodeRef {
-  return node instanceof Node || isForwardRefNode(node);
-}
-
-/** @hidden */
-export function isNodeRefLike(node: unknown): node is NodeRefLike {
-  return typeof node === 'string' || isEdgeTarget(node);
-}
-
-/** @hidden */
-export function isNodeRefGroupLike(target: NodeRefLike | NodeRefGroupLike): target is NodeRefGroupLike {
-  return Array.isArray(target) && target.every(isNodeRefLike);
 }
